@@ -23,6 +23,7 @@ import (
 	"encoding/base64"
 	stdjson "encoding/json"
 	"fmt"
+	"math/big"
 	"testing"
 
 	"github.com/AlekSi/pointer"
@@ -34,7 +35,7 @@ import (
 )
 
 func TestClient_GetAccountInfo(t *testing.T) {
-	responseBody := `{"context":{"slot":83986105},"value":{"data":["dGVzdA==","base64"],"executable":true,"lamports":999999,"owner":"11111111111111111111111111111111","rentEpoch":207}}`
+	responseBody := `{"context":{"slot":83986105},"value":{"data":["dGVzdA==","base64"],"executable":true,"lamports":999999,"owner":"11111111111111111111111111111111","rentEpoch":18446744073709551615}}`
 	server, closer := mockJSONRPC(t, stdjson.RawMessage(wrapIntoRPC(responseBody)))
 	defer closer()
 	client := New(server.URL)
@@ -44,9 +45,14 @@ func TestClient_GetAccountInfo(t *testing.T) {
 	out, err := client.GetAccountInfo(context.Background(), pubKey)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getAccountInfo",
 			"params": []interface{}{
@@ -56,9 +62,10 @@ func TestClient_GetAccountInfo(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
+	rentEpoch, _ := new(big.Int).SetString("18446744073709551615", 10)
 	assert.Equal(t,
 		&GetAccountInfoResult{
 			RPCContext: RPCContext{
@@ -75,7 +82,7 @@ func TestClient_GetAccountInfo(t *testing.T) {
 					},
 				},
 				Executable: true,
-				RentEpoch:  207,
+				RentEpoch:  rentEpoch,
 			},
 		}, out)
 }
@@ -109,9 +116,14 @@ func TestClient_GetAccountInfoWithOpts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getAccountInfo",
 			"params": []interface{}{
@@ -127,12 +139,12 @@ func TestClient_GetAccountInfoWithOpts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 }
 
 func TestClient_GetConfirmedSignaturesForAddress2(t *testing.T) {
-	server, closer := mockJSONRPC(t, stdjson.RawMessage(`{"jsonrpc":"2.0","result":[{"err":null,"memo":null,"signature":"mgw5vw4tnbou1wVStKckVcVncbpRwfZPcMNbVBoigbSPXBMa3857CNzhwoCkRzM5K7nG32wcbpVJDHttQeBRaHB","slot":1}],"id":0}`))
+	server, closer := mockJSONRPC(t, stdjson.RawMessage(`{"jsonrpc":"2.0","result":[{"err":null,"memo":null,"signature":"mgw5vw4tnbou1wVStKckVcVncbpRwfZPcMNbVBoigbSPXBMa3857CNzhwoCkRzM5K7nG32wcbpVJDHttQeBRaHB","slot":1}],"id":null}`))
 	defer closer()
 	client := New(server.URL)
 
@@ -141,9 +153,14 @@ func TestClient_GetConfirmedSignaturesForAddress2(t *testing.T) {
 	out, err := client.GetConfirmedSignaturesForAddress2(context.Background(), account, &GetConfirmedSignaturesForAddress2Opts{Limit: &limit})
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getConfirmedSignaturesForAddress2",
 			"params": []interface{}{
@@ -151,7 +168,7 @@ func TestClient_GetConfirmedSignaturesForAddress2(t *testing.T) {
 				map[string]interface{}{"limit": float64(1)},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := []*TransactionSignature{
@@ -162,7 +179,7 @@ func TestClient_GetConfirmedSignaturesForAddress2(t *testing.T) {
 }
 
 func TestClient_GetConfirmedTransaction(t *testing.T) {
-	server, closer := mockJSONRPC(t, stdjson.RawMessage(`{"jsonrpc":"2.0","result":{"meta":{"err":null,"fee":5000,"innerInstructions":[],"logMessages":[],"postBalances":[],"preBalances":[],"status":{"Ok":null}},"slot":48291656,"transaction":["AcpmPgtaSCzI2vuOUXduljmnoc1zIqMETzEJ8zmF+\/yy2AABHMNonpVleveVw4a4Fo7LUDWtxo2FkyzFr2x9DQIBAAMB47aX3y9Dfp+\/ycSDXt0Ph3TfZQBqPSXMQYToKtUtr5kNhniVeV7Las6qkeV8d0rksxV9de0GF7p4nzQUVEnrWwEEBAECAwAEdGVzdA==","base64"]},"id":0}`))
+	server, closer := mockJSONRPC(t, stdjson.RawMessage(`{"jsonrpc":"2.0","result":{"meta":{"err":null,"fee":5000,"innerInstructions":[],"logMessages":[],"postBalances":[],"preBalances":[],"status":{"Ok":null}},"slot":48291656,"transaction":["AcpmPgtaSCzI2vuOUXduljmnoc1zIqMETzEJ8zmF+\/yy2AABHMNonpVleveVw4a4Fo7LUDWtxo2FkyzFr2x9DQIBAAMB47aX3y9Dfp+\/ycSDXt0Ph3TfZQBqPSXMQYToKtUtr5kNhniVeV7Las6qkeV8d0rksxV9de0GF7p4nzQUVEnrWwEEBAECAwAEdGVzdA==","base64"]},"id":null}`))
 	defer closer()
 	client := New(server.URL)
 
@@ -172,9 +189,14 @@ func TestClient_GetConfirmedTransaction(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getConfirmedTransaction",
 			"params": []interface{}{
@@ -182,7 +204,7 @@ func TestClient_GetConfirmedTransaction(t *testing.T) {
 				"json",
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	signature, err := solana.SignatureFromBase58("53hoZ98EsCMA6L63GWM65M3Bd3WqA4LxD8bcJkbKoKWhbJFqX9M1WZ4fSjt8bYyZn21NwNnV2A25zirBni9Qk6LR")
@@ -265,9 +287,14 @@ func TestClient_GetRecentBlockhash(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getRecentBlockhash",
 			"params": []interface{}{
@@ -276,7 +303,7 @@ func TestClient_GetRecentBlockhash(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -302,9 +329,14 @@ func TestClient_GetBalance(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBalance",
 			"params": []interface{}{
@@ -314,7 +346,7 @@ func TestClient_GetBalance(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	assert.Equal(t,
@@ -340,9 +372,14 @@ func TestClient_GetBlock(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlock",
 			"params": []interface{}{
@@ -352,7 +389,7 @@ func TestClient_GetBlock(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	// TODO:
@@ -507,9 +544,14 @@ func TestClient_GetBlockWithOpts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlock",
 			"params": []interface{}{
@@ -523,7 +565,7 @@ func TestClient_GetBlockWithOpts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	// TODO:
@@ -542,9 +584,14 @@ func TestClient_GetBlockHeight(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlockHeight",
 			"params": []interface{}{
@@ -553,7 +600,7 @@ func TestClient_GetBlockHeight(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -574,14 +621,19 @@ func TestClient_GetBlockProduction(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlockProduction",
 			"params":  []interface{}{},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -615,9 +667,14 @@ func TestClient_GetBlockProductionWithOpts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlockProduction",
 			"params": []interface{}{
@@ -631,7 +688,7 @@ func TestClient_GetBlockProductionWithOpts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 }
 
@@ -649,16 +706,21 @@ func TestClient_GetBlockCommitment(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlockCommitment",
 			"params": []interface{}{
 				float64(block),
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := map[string]interface{}{
@@ -720,9 +782,14 @@ func TestClient_GetBlocks(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlocks",
 			"params": []interface{}{
@@ -733,7 +800,7 @@ func TestClient_GetBlocks(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -759,9 +826,14 @@ func TestClient_GetBlocksWithLimit(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlocksWithLimit",
 			"params": []interface{}{
@@ -772,7 +844,7 @@ func TestClient_GetBlocksWithLimit(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -795,16 +867,21 @@ func TestClient_GetBlockTime(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getBlockTime",
 			"params": []interface{}{
 				float64(block),
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -815,7 +892,7 @@ func TestClient_GetBlockTime(t *testing.T) {
 }
 
 func TestClient_GetClusterNodes(t *testing.T) {
-	responseBody := `[{"featureSet":743297851,"gossip":"162.55.111.250:8001","pubkey":"DMeohMfD3JzmYZA34jL9iiTXp5N7tpAR3rAoXMygdH3U","rpc":"135.181.114.15:8005","shredVersion":18122,"tpu":"162.55.111.250:8004","version":"1.7.3"},{"featureSet":743297851,"gossip":"136.243.131.82:8000","pubkey":"59TSbYfnbb4zx4xf54ApjE8fJRhwzTiSjh9vdHfgyg1U","rpc":"136.243.131.82:8899","shredVersion":18122,"tpu":"136.243.131.82:8003","version":"1.7.3"},{"featureSet":743297851,"gossip":"135.181.114.15:8001","pubkey":"7vu7Q2d4uu9V4xnySHXieeyWvoNh37321kqTd2ATuoj6","rpc":"135.181.114.15:8005","shredVersion":18122,"tpu":"135.181.114.15:8006","version":"1.7.3"}]`
+	responseBody := `[{"featureSet":3580551090,"gossip":"34.147.255.155:8000","pubkey":"hyp3Eo67t6FgeuWg5Qxbeme8NPXJPXXdKT4iJ4DsLf2","pubsub":"34.147.255.155:8900","rpc":"34.147.255.155:8899","shredVersion":50093,"tpu":"34.147.255.155:8009","tpuQuic":"34.147.255.155:8015","version":"1.17.22"},{"featureSet":3746964731,"gossip":"162.19.222.39:8001","pubkey":"EvnRmnMrd69kFdbLMxWkTn1icZ7DCceRhvmb2SJXqDo4","pubsub":"162.19.222.39:8900","rpc":"162.19.222.39:8899","shredVersion":50093,"tpu":"208.91.106.87:8005","tpuQuic":"208.91.106.87:8011","version":"1.17.27"},{"featureSet":3746964731,"gossip":"205.209.104.74:8000","pubkey":"J87afqF2bDQQLTQpks4SdF7hXPr96SPTdJ28UJXXWr9N","pubsub":"205.209.104.74:8900","rpc":"205.209.104.74:8899","shredVersion":50093,"tpu":"205.209.104.74:8003","tpuQuic":"205.209.104.74:8009","version":"1.17.27"}]`
 	server, closer := mockJSONRPC(t, stdjson.RawMessage(wrapIntoRPC(responseBody)))
 	defer closer()
 	client := New(server.URL)
@@ -825,13 +902,18 @@ func TestClient_GetClusterNodes(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getClusterNodes",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -853,9 +935,14 @@ func TestClient_GetEpochInfo(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getEpochInfo",
 			"params": []interface{}{
@@ -864,7 +951,7 @@ func TestClient_GetEpochInfo(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := map[string]interface{}{
@@ -892,13 +979,18 @@ func TestClient_GetEpochSchedule(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getEpochSchedule",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -921,9 +1013,14 @@ func TestClient_GetFeeCalculatorForBlockhash(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getFeeCalculatorForBlockhash",
 			"params": []interface{}{
@@ -933,7 +1030,7 @@ func TestClient_GetFeeCalculatorForBlockhash(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -954,13 +1051,18 @@ func TestClient_GetFeeRateGovernor(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getFeeRateGovernor",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -982,9 +1084,14 @@ func TestClient_GetFees(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getFees",
 			"params": []interface{}{
@@ -993,7 +1100,7 @@ func TestClient_GetFees(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1014,13 +1121,18 @@ func TestClient_GetFirstAvailableBlock(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getFirstAvailableBlock",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1041,13 +1153,18 @@ func TestClient_GetGenesisHash(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getGenesisHash",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1068,13 +1185,18 @@ func TestClient_GetHealth(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getHealth",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1095,13 +1217,18 @@ func TestClient_GetIdentity(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getIdentity",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1123,9 +1250,14 @@ func TestClient_GetInflationGovernor(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getInflationGovernor",
 			"params": []interface{}{
@@ -1134,7 +1266,7 @@ func TestClient_GetInflationGovernor(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1155,13 +1287,18 @@ func TestClient_GetInflationRate(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getInflationRate",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1196,9 +1333,14 @@ func TestClient_GetInflationReward(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getInflationReward",
 			"params": []interface{}{
@@ -1211,7 +1353,7 @@ func TestClient_GetInflationReward(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1235,9 +1377,14 @@ func TestClient_GetLargestAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getLargestAccounts",
 			"params": []interface{}{
@@ -1247,7 +1394,7 @@ func TestClient_GetLargestAccounts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := &GetLargestAccountsResult{
@@ -1364,9 +1511,14 @@ func TestClient_GetLeaderSchedule(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getLeaderSchedule",
 			"params": []interface{}{
@@ -1377,7 +1529,7 @@ func TestClient_GetLeaderSchedule(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1398,13 +1550,18 @@ func TestClient_GetMaxRetransmitSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getMaxRetransmitSlot",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1425,13 +1582,18 @@ func TestClient_GetMaxShredInsertSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getMaxShredInsertSlot",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1455,9 +1617,14 @@ func TestClient_GetMinimumBalanceForRentExemption(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getMinimumBalanceForRentExemption",
 			"params": []interface{}{
@@ -1467,7 +1634,7 @@ func TestClient_GetMinimumBalanceForRentExemption(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1491,16 +1658,21 @@ func TestClient_GetMultipleAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getMultipleAccounts",
 			"params": []interface{}{
 				[]interface{}{pubkeyString},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := &GetMultipleAccountsResult{
@@ -1521,7 +1693,7 @@ func TestClient_GetMultipleAccounts(t *testing.T) {
 					rawDataEncoding: solana.EncodingBase64,
 				},
 				Executable: true,
-				RentEpoch:  207,
+				RentEpoch:  big.NewInt(207),
 			},
 		},
 	}
@@ -1564,9 +1736,14 @@ func TestClient_GetProgramAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getProgramAccounts",
 			"params": []interface{}{
@@ -1589,7 +1766,7 @@ func TestClient_GetProgramAccounts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := GetProgramAccountsResult{
@@ -1606,7 +1783,7 @@ func TestClient_GetProgramAccounts(t *testing.T) {
 					rawDataEncoding: solana.EncodingBase64,
 				},
 				Executable: true,
-				RentEpoch:  206,
+				RentEpoch:  big.NewInt(206),
 			},
 		},
 	}
@@ -1627,16 +1804,21 @@ func TestClient_GetRecentPerformanceSamples(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getRecentPerformanceSamples",
 			"params": []interface{}{
 				float64(limit),
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1657,13 +1839,18 @@ func TestClient_GetSnapshotSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSnapshotSlot",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1701,9 +1888,14 @@ func TestClient_GetSignaturesForAddress(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSignaturesForAddress",
 			"params": []interface{}{
@@ -1717,7 +1909,7 @@ func TestClient_GetSignaturesForAddress(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1743,9 +1935,14 @@ func TestClient_GetSignatureStatuses(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSignatureStatuses",
 			"params": []interface{}{
@@ -1758,7 +1955,7 @@ func TestClient_GetSignatureStatuses(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1780,9 +1977,14 @@ func TestClient_GetSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSlot",
 			"params": []interface{}{
@@ -1791,7 +1993,7 @@ func TestClient_GetSlot(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1813,9 +2015,14 @@ func TestClient_GetSlotLeader(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSlotLeader",
 			"params": []interface{}{
@@ -1824,7 +2031,7 @@ func TestClient_GetSlotLeader(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1849,9 +2056,14 @@ func TestClient_GetSlotLeaders(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSlotLeaders",
 			"params": []interface{}{
@@ -1859,7 +2071,7 @@ func TestClient_GetSlotLeaders(t *testing.T) {
 				float64(limit),
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1878,9 +2090,14 @@ func TestClient_GetSupply(t *testing.T) {
 	out, err := client.GetSupply(context.Background(), CommitmentFinalized)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSupply",
 			"params": []interface{}{
@@ -1890,7 +2107,7 @@ func TestClient_GetSupply(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1915,9 +2132,14 @@ func TestClient_GetSupply_CommitmentMax(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSupply",
 			"params": []interface{}{
@@ -1927,7 +2149,7 @@ func TestClient_GetSupply_CommitmentMax(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1953,9 +2175,14 @@ func TestClient_GetSupply_ExcludeNonCirculatingAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getSupply",
 			"params": []interface{}{
@@ -1965,7 +2192,7 @@ func TestClient_GetSupply_ExcludeNonCirculatingAccounts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -1991,9 +2218,14 @@ func TestClient_GetTokenLargestAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTokenLargestAccounts",
 			"params": []interface{}{
@@ -2003,7 +2235,7 @@ func TestClient_GetTokenLargestAccounts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2029,9 +2261,14 @@ func TestClient_GetTokenSupply(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTokenSupply",
 			"params": []interface{}{
@@ -2041,7 +2278,7 @@ func TestClient_GetTokenSupply(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2072,9 +2309,14 @@ func TestClient_GetTransaction(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTransaction",
 			"params": []interface{}{
@@ -2086,7 +2328,7 @@ func TestClient_GetTransaction(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	blockTimeSeconds := int64(1624821990)
@@ -2180,9 +2422,14 @@ func TestClient_GetParsedTransaction(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTransaction",
 			"params": []interface{}{
@@ -2193,7 +2440,7 @@ func TestClient_GetParsedTransaction(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	assert.Equal(t, uint64(2), out.Meta.InnerInstructions[0].Index)
@@ -2228,9 +2475,14 @@ func TestClient_GetTransactionCount(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTransactionCount",
 			"params": []interface{}{
@@ -2239,7 +2491,7 @@ func TestClient_GetTransactionCount(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2260,13 +2512,18 @@ func TestClient_GetVersion(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getVersion",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2292,9 +2549,14 @@ func TestClient_GetVoteAccounts(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getVoteAccounts",
 			"params": []interface{}{
@@ -2304,7 +2566,7 @@ func TestClient_GetVoteAccounts(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2325,13 +2587,18 @@ func TestClient_MinimumLedgerSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "minimumLedgerSlot",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2359,9 +2626,14 @@ func TestClient_RequestAirdrop(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "requestAirdrop",
 			"params": []interface{}{
@@ -2372,7 +2644,7 @@ func TestClient_RequestAirdrop(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2400,9 +2672,14 @@ func TestClient_GetStakeActivation(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getStakeActivation",
 			"params": []interface{}{
@@ -2413,7 +2690,7 @@ func TestClient_GetStakeActivation(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2439,9 +2716,14 @@ func TestClient_GetTokenAccountBalance(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTokenAccountBalance",
 			"params": []interface{}{
@@ -2451,7 +2733,7 @@ func TestClient_GetTokenAccountBalance(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2462,7 +2744,7 @@ func TestClient_GetTokenAccountBalance(t *testing.T) {
 }
 
 func TestClient_GetTokenAccountsByDelegate(t *testing.T) {
-	responseBody := `{"context":{"slot":1114},"value":[{"account":{"data":{"program":"spl-token","parsed":{"accountType":"account","info":{"tokenAmount":{"amount":"1","decimals":1,"uiAmount":0.1,"uiAmountString":"0.1"},"delegate":"4Nd1mBQtrMJVYVfKf2PJy9NZUZdTAsp7D4xWLs4gDB4T","delegatedAmount":1,"isInitialized":true,"isNative":false,"mint":"3wyAj7Rt1TWVPZVteFJPLa26JmLvdb1CAKEFZm3NY75E","owner":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}}},"executable":false,"lamports":1726080,"owner":"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA","rentEpoch":4},"pubkey":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}]}`
+	responseBody := `{"context":{"slot":1114},"value":[{"account":{"data":{"program":"spl-token","parsed":{"accountType":"account","info":{"tokenAmount":{"amount":"1","decimals":1,"uiAmount":0.1,"uiAmountString":"0.1"},"delegate":"4Nd1mBQtrMJVYVfKf2PJy9NZUZdTAsp7D4xWLs4gDB4T","delegatedAmount":1,"isInitialized":true,"isNative":false,"mint":"3wyAj7Rt1TWVPZVteFJPLa26JmLvdb1CAKEFZm3NY75E","owner":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}}},"executable":false,"lamports":1726080,"owner":"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA","rentEpoch":4,"space":0},"pubkey":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}]}`
 	server, closer := mockJSONRPC(t, stdjson.RawMessage(wrapIntoRPC(responseBody)))
 	defer closer()
 	client := New(server.URL)
@@ -2486,9 +2768,14 @@ func TestClient_GetTokenAccountsByDelegate(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTokenAccountsByDelegate",
 			"params": []interface{}{
@@ -2502,7 +2789,7 @@ func TestClient_GetTokenAccountsByDelegate(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2513,7 +2800,7 @@ func TestClient_GetTokenAccountsByDelegate(t *testing.T) {
 }
 
 func TestClient_GetTokenAccountsByOwner(t *testing.T) {
-	responseBody := `{"context":{"slot":1114},"value":[{"account":{"data":{"program":"spl-token","parsed":{"accountType":"account","info":{"tokenAmount":{"amount":"1","decimals":1,"uiAmount":0.1,"uiAmountString":"0.1"},"delegate":null,"delegatedAmount":1,"isInitialized":true,"isNative":false,"mint":"3wyAj7Rt1TWVPZVteFJPLa26JmLvdb1CAKEFZm3NY75E","owner":"4Qkev8aNZcqFNSRhQzwyLMFSsi94jHqE8WNVTJzTP99F"}}},"executable":false,"lamports":1726080,"owner":"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA","rentEpoch":4},"pubkey":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}]}`
+	responseBody := `{"context":{"slot":1114},"value":[{"account":{"data":{"program":"spl-token","parsed":{"accountType":"account","info":{"tokenAmount":{"amount":"1","decimals":1,"uiAmount":0.1,"uiAmountString":"0.1"},"delegate":null,"delegatedAmount":1,"isInitialized":true,"isNative":false,"mint":"3wyAj7Rt1TWVPZVteFJPLa26JmLvdb1CAKEFZm3NY75E","owner":"4Qkev8aNZcqFNSRhQzwyLMFSsi94jHqE8WNVTJzTP99F"}}},"executable":false,"lamports":1726080,"owner":"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA","rentEpoch":4,"space":0},"pubkey":"CnPoSPKXu7wJqxe59Fs72tkBeALovhsCxYeFwPCQH9TD"}]}`
 	server, closer := mockJSONRPC(t, stdjson.RawMessage(wrapIntoRPC(responseBody)))
 	defer closer()
 	client := New(server.URL)
@@ -2537,9 +2824,14 @@ func TestClient_GetTokenAccountsByOwner(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getTokenAccountsByOwner",
 			"params": []interface{}{
@@ -2553,7 +2845,7 @@ func TestClient_GetTokenAccountsByOwner(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2644,9 +2936,14 @@ func TestClient_IsBlockhashValid(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "isBlockhashValid",
 			"params": []interface{}{
@@ -2656,7 +2953,7 @@ func TestClient_IsBlockhashValid(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	assert.Equal(t,
@@ -2685,9 +2982,14 @@ func TestClient_GetFeeForMessage(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getFeeForMessage",
 			"params": []interface{}{
@@ -2697,7 +2999,7 @@ func TestClient_GetFeeForMessage(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2718,13 +3020,18 @@ func TestClient_GetHighestSnapshotSlot(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getHighestSnapshotSlot",
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2746,9 +3053,14 @@ func TestClient_GetLatestBlockhash(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getLatestBlockhash",
 			"params": []interface{}{
@@ -2757,7 +3069,7 @@ func TestClient_GetLatestBlockhash(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
@@ -2785,9 +3097,14 @@ func TestClient_GetRecentPrioritizationFees(t *testing.T) {
 	)
 	require.NoError(t, err)
 
+	// the ID is random, so we can't assert it; let's check that it is set, and then remove it
+	reqBody := server.RequestBody(t)
+	assert.NotNil(t, reqBody["id"])
+	reqBody["id"] = any(nil)
+
 	assert.Equal(t,
 		map[string]interface{}{
-			"id":      float64(0),
+			"id":      any(nil),
 			"jsonrpc": "2.0",
 			"method":  "getRecentPrioritizationFees",
 			"params": []interface{}{
@@ -2797,7 +3114,7 @@ func TestClient_GetRecentPrioritizationFees(t *testing.T) {
 				},
 			},
 		},
-		server.RequestBody(t),
+		reqBody,
 	)
 
 	expected := mustJSONToInterface([]byte(responseBody))
