@@ -99,7 +99,7 @@ func (cl *Client) GetAccountInfoWithOpts(
 	if err != nil {
 		return nil, err
 	}
-	if out.Value == nil {
+	if out == nil || out.Value == nil {
 		return nil, ErrNotFound
 	}
 	return out, nil
