@@ -12,6 +12,7 @@ import (
 	"testing"
 
 	. "github.com/onsi/gomega"
+	"github.com/stretchr/testify/require"
 )
 
 // needed to retrieve requests that arrived at httpServer for further investigation
@@ -72,96 +73,98 @@ func TestRpcClient_Call(t *testing.T) {
 		Ingredients: []string{"rum", "cola"},
 	}
 
+	useFixedID = true
+
 	rpcClient.Call(context.Background(), "missingParam")
-	Expect((<-requestChan).body).To(Equal(`{"method":"missingParam","id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"missingParam","id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "nullParam", nil)
-	Expect((<-requestChan).body).To(Equal(`{"method":"nullParam","params":[null],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"nullParam","params":[null],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "nullParams", nil, nil)
-	Expect((<-requestChan).body).To(Equal(`{"method":"nullParams","params":[null,null],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"nullParams","params":[null,null],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "emptyParams", []interface{}{})
-	Expect((<-requestChan).body).To(Equal(`{"method":"emptyParams","params":[],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"emptyParams","params":[],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "emptyAnyParams", []string{})
-	Expect((<-requestChan).body).To(Equal(`{"method":"emptyAnyParams","params":[],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"emptyAnyParams","params":[],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "emptyObject", struct{}{})
-	Expect((<-requestChan).body).To(Equal(`{"method":"emptyObject","params":{},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"emptyObject","params":{},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "emptyObjectList", []struct{}{{}, {}})
-	Expect((<-requestChan).body).To(Equal(`{"method":"emptyObjectList","params":[{},{}],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"emptyObjectList","params":[{},{}],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "boolParam", true)
-	Expect((<-requestChan).body).To(Equal(`{"method":"boolParam","params":[true],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"boolParam","params":[true],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "boolParams", true, false, true)
-	Expect((<-requestChan).body).To(Equal(`{"method":"boolParams","params":[true,false,true],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"boolParams","params":[true,false,true],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "stringParam", "Alex")
-	Expect((<-requestChan).body).To(Equal(`{"method":"stringParam","params":["Alex"],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"stringParam","params":["Alex"],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "stringParams", "JSON", "RPC")
-	Expect((<-requestChan).body).To(Equal(`{"method":"stringParams","params":["JSON","RPC"],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"stringParams","params":["JSON","RPC"],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "numberParam", 123)
-	Expect((<-requestChan).body).To(Equal(`{"method":"numberParam","params":[123],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"numberParam","params":[123],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "numberParams", 123, 321)
-	Expect((<-requestChan).body).To(Equal(`{"method":"numberParams","params":[123,321],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"numberParams","params":[123,321],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "floatParam", 1.23)
-	Expect((<-requestChan).body).To(Equal(`{"method":"floatParam","params":[1.23],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"floatParam","params":[1.23],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "floatParams", 1.23, 3.21)
-	Expect((<-requestChan).body).To(Equal(`{"method":"floatParams","params":[1.23,3.21],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"floatParams","params":[1.23,3.21],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "manyParams", "Alex", 35, true, nil, 2.34)
-	Expect((<-requestChan).body).To(Equal(`{"method":"manyParams","params":["Alex",35,true,null,2.34],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"manyParams","params":["Alex",35,true,null,2.34],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "emptyMissingPublicFieldObject", struct{ name string }{name: "Alex"})
-	Expect((<-requestChan).body).To(Equal(`{"method":"emptyMissingPublicFieldObject","params":{},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"emptyMissingPublicFieldObject","params":{},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "singleStruct", person)
-	Expect((<-requestChan).body).To(Equal(`{"method":"singleStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"singleStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "singlePointerToStruct", &person)
-	Expect((<-requestChan).body).To(Equal(`{"method":"singlePointerToStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"singlePointerToStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":1,"jsonrpc":"2.0"}`))
 
 	pp := &person
 	rpcClient.Call(context.Background(), "doublePointerStruct", &pp)
-	Expect((<-requestChan).body).To(Equal(`{"method":"doublePointerStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"doublePointerStruct","params":{"name":"Alex","age":35,"country":"Germany"},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "multipleStructs", person, &drink)
-	Expect((<-requestChan).body).To(Equal(`{"method":"multipleStructs","params":[{"name":"Alex","age":35,"country":"Germany"},{"name":"Cuba Libre","ingredients":["rum","cola"]}],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"multipleStructs","params":[{"name":"Alex","age":35,"country":"Germany"},{"name":"Cuba Libre","ingredients":["rum","cola"]}],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "singleStructInArray", []interface{}{person})
-	Expect((<-requestChan).body).To(Equal(`{"method":"singleStructInArray","params":[{"name":"Alex","age":35,"country":"Germany"}],"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"singleStructInArray","params":[{"name":"Alex","age":35,"country":"Germany"}],"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "namedParameters", map[string]interface{}{
 		"name": "Alex",
 		"age":  35,
 	})
-	Expect((<-requestChan).body).To(Equal(`{"method":"namedParameters","params":{"age":35,"name":"Alex"},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"namedParameters","params":{"age":35,"name":"Alex"},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "anonymousStructNoTags", struct {
 		Name string
 		Age  int
 	}{"Alex", 33})
-	Expect((<-requestChan).body).To(Equal(`{"method":"anonymousStructNoTags","params":{"Name":"Alex","Age":33},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"anonymousStructNoTags","params":{"Name":"Alex","Age":33},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "anonymousStructWithTags", struct {
 		Name string `json:"name"`
 		Age  int    `json:"age"`
 	}{"Alex", 33})
-	Expect((<-requestChan).body).To(Equal(`{"method":"anonymousStructWithTags","params":{"name":"Alex","age":33},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"anonymousStructWithTags","params":{"name":"Alex","age":33},"id":1,"jsonrpc":"2.0"}`))
 
 	rpcClient.Call(context.Background(), "structWithNullField", struct {
 		Name    string  `json:"name"`
 		Address *string `json:"address"`
 	}{"Alex", nil})
-	Expect((<-requestChan).body).To(Equal(`{"method":"structWithNullField","params":{"name":"Alex","address":null},"id":0,"jsonrpc":"2.0"}`))
+	Expect((<-requestChan).body).To(Equal(`{"method":"structWithNullField","params":{"name":"Alex","address":null},"id":1,"jsonrpc":"2.0"}`))
 }
 
 func TestRpcClient_CallBatch(t *testing.T) {
@@ -646,14 +649,14 @@ func TestRpcBatchJsonResponseStruct(t *testing.T) {
 	Expect(res).To(BeNil())*/
 
 	// result string is ok
-	responseBody = `[{"result": "ok","id":0}]`
+	responseBody = `[{"result": "ok","id":1}]`
 	res, err = rpcClient.CallBatch(context.Background(), RPCRequests{
 		NewRequest("something", 1, 2, 3),
 	})
 	<-requestChan
 	Expect(err).To(BeNil())
 	Expect(res[0].Result).To(Equal(stdjson.RawMessage([]byte(strconv.Quote("ok")))))
-	Expect(res[0].ID).To(Equal(0))
+	Expect(res[0].ID).To(Equal(stdjson.Number("1")))
 
 	// result with error null is ok
 	responseBody = `[{"result": "ok", "error": null}]`
@@ -719,7 +722,7 @@ func TestRpcBatchJsonResponseStruct(t *testing.T) {
 	// check results
 
 	var p *Person
-	responseBody = `[{"id":0, "result": {"name": "Alex", "age": 35}}, {"id":2, "result": {"name": "Lena", "age": 2}}]`
+	responseBody = `[{"id":1, "result": {"name": "Alex", "age": 35}}, {"id":2, "result": {"name": "Lena", "age": 2}}]`
 	res, err = rpcClient.CallBatch(context.Background(), RPCRequests{
 		NewRequest("something", 1, 2, 3),
 	})
@@ -728,16 +731,18 @@ func TestRpcBatchJsonResponseStruct(t *testing.T) {
 	Expect(err).To(BeNil())
 
 	Expect(res[0].Error).To(BeNil())
-	Expect(res[0].ID).To(Equal(0))
+	Expect(res[0].ID).To(Equal(stdjson.Number("1")))
 
 	Expect(res[1].Error).To(BeNil())
-	Expect(res[1].ID).To(Equal(2))
+	Expect(res[1].ID).To(Equal(stdjson.Number("2")))
 
 	err = res[0].GetObject(&p)
+	require.NoError(t, err)
 	Expect(p.Name).To(Equal("Alex"))
 	Expect(p.Age).To(Equal(35))
 
 	err = res[1].GetObject(&p)
+	require.NoError(t, err)
 	Expect(p.Name).To(Equal("Lena"))
 	Expect(p.Age).To(Equal(2))
 
@@ -778,8 +783,10 @@ func TestRpcBatchJsonResponseStruct(t *testing.T) {
 	resMap := res.AsMap()
 
 	var int1 int64
+	require.NotNil(t, resMap[1])
 	resMap[1].GetObject(&int1)
 	var int123 int64
+	require.NotNil(t, resMap[123])
 	resMap[123].GetObject(&int123)
 	Expect(int1).To(Equal(int64(1)))
 	Expect(int123).To(Equal(int64(123)))
@@ -931,7 +938,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 	rpcClient := NewClient(httpServer.URL)
 
 	i := 0
-	responseBody = `{"result":3,"id":0,"jsonrpc":"2.0"}`
+	responseBody = `{"result":3,"id":1,"jsonrpc":"2.0"}`
 	err := rpcClient.CallFor(context.Background(), &i, "something", 1, 2, 3)
 	<-requestChan
 	Expect(err).To(BeNil())
@@ -939,7 +946,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 
 	/*
 		i = 3
-		responseBody = `{"result":null,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":null,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&i, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -947,20 +954,20 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(i).To(Equal(3))
 
 		var pi *int
-		responseBody = `{"result":4,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":4,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(pi, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).NotTo(BeNil())
 		Expect(pi).To(BeNil())
 
-		responseBody = `{"result":4,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":4,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&pi, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
 		Expect(*pi).To(Equal(4))
 
 		*pi = 3
-		responseBody = `{"result":null,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":null,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&pi, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -968,7 +975,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(pi).To(BeNil())
 
 		p := &Person{}
-		responseBody = `{"result":null,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":null,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(p, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -976,7 +983,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p).NotTo(BeNil())
 
 		var p2 *Person
-		responseBody = `{"result":null,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":null,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(p2, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).NotTo(BeNil())
@@ -984,7 +991,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p2).To(BeNil())
 
 		p3 := Person{}
-		responseBody = `{"result":null,"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":null,"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&p3, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -992,7 +999,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p).NotTo(BeNil())
 
 		p = &Person{Age: 35}
-		responseBody = `{"result":{"name":"Alex"},"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":{"name":"Alex"},"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(p, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -1001,7 +1008,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p.Age).To(Equal(35))
 
 		p2 = nil
-		responseBody = `{"result":{"name":"Alex"},"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":{"name":"Alex"},"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(p2, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).NotTo(BeNil())
@@ -1009,7 +1016,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p2).To(BeNil())
 
 		p2 = nil
-		responseBody = `{"result":{"name":"Alex"},"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":{"name":"Alex"},"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&p2, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -1018,7 +1025,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p2.Name).To(Equal("Alex"))
 
 		p3 = Person{Age: 35}
-		responseBody = `{"result":{"name":"Alex"},"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":{"name":"Alex"},"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&p3, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -1027,7 +1034,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p.Age).To(Equal(35))
 
 		p3 = Person{Age: 35}
-		responseBody = `{"result":{"name":"Alex"},"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":{"name":"Alex"},"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&p3, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
@@ -1036,7 +1043,7 @@ func TestRpcClient_CallFor(t *testing.T) {
 		Expect(p.Age).To(Equal(35))
 
 		var intArray []int
-		responseBody = `{"result":[1, 2, 3],"id":0,"jsonrpc":"2.0"}`
+		responseBody = `{"result":[1, 2, 3],"id":1,"jsonrpc":"2.0"}`
 		err = rpcClient.CallFor(&intArray, "something", 1, 2, 3)
 		<-requestChan
 		Expect(err).To(BeNil())
