@@ -45,14 +45,14 @@ func Test_AccountSubscribe(t *testing.T) {
 	sub, err := c.AccountSubscribe(accountID, "")
 	require.NoError(t, err)
 
-	data, err := sub.Recv()
+	data, err := sub.Recv(context.Background())
 	if err != nil {
 		fmt.Println("receive an error: ", err)
 		return
 	}
 	text.NewEncoder(os.Stdout).Encode(data, nil)
-	fmt.Println("OpenOrders: ", data.Value.Account.Owner)
-	fmt.Println("data: ", data.Value.Account.Data)
+	fmt.Println("OpenOrders: ", data.Value.Owner)
+	fmt.Println("data: ", data.Value.Data)
 	return
 }
 
@@ -95,7 +95,7 @@ func Test_AccountSubscribeWithHttpHeader(t *testing.T) {
 		sub.Unsubscribe()
 	}(sub)
 
-	data, err := sub.Recv()
+	data, err := sub.Recv(context.Background())
 	if err != nil {
 		t.Errorf("Received an error: %v", err)
 	}
@@ -107,8 +107,8 @@ func Test_AccountSubscribeWithHttpHeader(t *testing.T) {
 		t.Errorf("encoding error: %v", err)
 	}
 
-	t.Log("OpenOrders: ", data.Value.Account.Owner)
-	t.Log("data: ", data.Value.Account.Data)
+	t.Log("OpenOrders: ", data.Value.Owner)
+	t.Log("data: ", data.Value.Data)
 }
 
 func Test_ProgramSubscribe(t *testing.T) {
@@ -127,7 +127,7 @@ func Test_ProgramSubscribe(t *testing.T) {
 	require.NoError(t, err)
 
 	for {
-		data, err := sub.Recv()
+		data, err := sub.Recv(context.Background())
 		if err != nil {
 			fmt.Println("receive an error: ", err)
 			return
@@ -148,7 +148,7 @@ func Test_SlotSubscribe(t *testing.T) {
 	sub, err := c.SlotSubscribe()
 	require.NoError(t, err)
 
-	data, err := sub.Recv()
+	data, err := sub.Recv(context.Background())
 	if err != nil {
 		fmt.Println("receive an error: ", err)
 		return
