@@ -24,6 +24,7 @@ import (
 )
 
 func main() {
+	ctx := context.Background()
 	client, err := ws.Connect(context.Background(), rpc.TestNet_WS)
 	if err != nil {
 		panic(err)
@@ -42,7 +43,7 @@ func main() {
 	defer sub.Unsubscribe()
 
 	for {
-		got, err := sub.Recv()
+		got, err := sub.Recv(ctx)
 		if err != nil {
 			panic(err)
 		}
