@@ -15,6 +15,8 @@
 package ws
 
 import (
+	"context"
+
 	"github.com/Arculus-Holdings-L-L-C/solana-go"
 	"github.com/Arculus-Holdings-L-L-C/solana-go/rpc"
 )
@@ -86,13 +88,35 @@ type ProgramSubscription struct {
 	sub *Subscription
 }
 
-func (sw *ProgramSubscription) Recv() (*ProgramResult, error) {
+func (sw *ProgramSubscription) Recv(ctx context.Context) (*ProgramResult, error) {
 	select {
-	case d := <-sw.sub.stream:
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case d, ok := <-sw.sub.stream:
+		if !ok {
+			return nil, ErrSubscriptionClosed
+		}
 		return d.(*ProgramResult), nil
 	case err := <-sw.sub.err:
 		return nil, err
 	}
+}
+
+func (sw *ProgramSubscription) Err() <-chan error {
+	return sw.sub.err
+}
+
+func (sw *ProgramSubscription) Response() <-chan *ProgramResult {
+	typedChan := make(chan *ProgramResult, 1)
+	go func(ch chan *ProgramResult) {
+		// TODO: will this subscription yield more than one result?
+		d, ok := <-sw.sub.stream
+		if !ok {
+			return
+		}
+		ch <- d.(*ProgramResult)
+	}(typedChan)
+	return typedChan
 }
 
 func (sw *ProgramSubscription) Unsubscribe() {
