@@ -14,6 +14,8 @@
 
 package ws
 
+import "context"
+
 type RootResult uint64
 
 // SignatureSubscribe subscribes to receive notification
@@ -42,13 +44,35 @@ type RootSubscription struct {
 	sub *Subscription
 }
 
-func (sw *RootSubscription) Recv() (*RootResult, error) {
+func (sw *RootSubscription) Recv(ctx context.Context) (*RootResult, error) {
 	select {
-	case d := <-sw.sub.stream:
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case d, ok := <-sw.sub.stream:
+		if !ok {
+			return nil, ErrSubscriptionClosed
+		}
 		return d.(*RootResult), nil
 	case err := <-sw.sub.err:
 		return nil, err
 	}
+}
+
+func (sw *RootSubscription) Err() <-chan error {
+	return sw.sub.err
+}
+
+func (sw *RootSubscription) Response() <-chan *RootResult {
+	typedChan := make(chan *RootResult, 1)
+	go func(ch chan *RootResult) {
+		// TODO: will this subscription yield more than one result?
+		d, ok := <-sw.sub.stream
+		if !ok {
+			return
+		}
+		ch <- d.(*RootResult)
+	}(typedChan)
+	return typedChan
 }
 
 func (sw *RootSubscription) Unsubscribe() {
