@@ -15,6 +15,7 @@
 package ws
 
 import (
+	"context"
 	"fmt"
 	"time"
 
@@ -67,9 +68,14 @@ type SignatureSubscription struct {
 	sub *Subscription
 }
 
-func (sw *SignatureSubscription) Recv() (*SignatureResult, error) {
+func (sw *SignatureSubscription) Recv(ctx context.Context) (*SignatureResult, error) {
 	select {
-	case d := <-sw.sub.stream:
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case d, ok := <-sw.sub.stream:
+		if !ok {
+			return nil, ErrSubscriptionClosed
+		}
 		return d.(*SignatureResult), nil
 	case err := <-sw.sub.err:
 		return nil, err
